+++ docs/misc/changelog.rst
@@ -14,8 +14,12 @@ Pre-Release 2.4.0a (WIP)
 - fixed bug related to shape of true_reward (@abhiskk)
 - fixed example code in documentation of tf_util:Function (@JohannesAck)
 - added learning rate schedule for SAC
+- fixed action probability for continuous actions with actor-critic models
+- added optional parameter to action_probability for likelihood calculation of given action being taken.
 - added more flexible custom LSTM policies
 - added auto entropy coefficient optimization for SAC
+- clip continuous actions at test time too for all algorithms (except SAC/DDPG where it is not needed)
+
 
 Release 2.3.0 (2018-12-05)
 --------------------------
@@ -210,3 +214,4 @@ Contributors (since v2.0.0):
 In random order...
 
 Thanks to @bjmuld @iambenzo @iandanforth @r7vme @brendenpetersen @huvar @abhiskk @JohannesAck
+@EliasHasle